Basic ACL Roles and rules admin screen done.

+++ src/RcmUser/Acl/Db/DoctrineAclRoleDataMapper.php
@@ -105,23 +105,12 @@ class DoctrineAclRoleDataMapper
      */
     public function fetchAll()
     {
-        //$roles = $this->getEntityManager()->getRepository(
-        //    $this->getEntityClass()
-        //)->findAll();
-
         $query = $this->getEntityManager()->createQuery(
             'SELECT role FROM ' . $this->getEntityClass() . ' role ' .
             'INDEX BY role.roleId'
         );
 
         $roles = $query->getResult();
-
-        if (empty($roles)) {
-
-            return new Result(null, Result::CODE_FAIL, 'Roles could not be found.');
-        }
-
-        //$roles = $this->prepareRoles($roles);
 
         return new Result($roles);
     }
@@ -138,15 +127,6 @@ class DoctrineAclRoleDataMapper
         $roles = $this->getEntityManager()->getRepository($this->getEntityClass())
             ->findOneBy(array('roleId' => $roleId));
 
-        if (empty($roles)) {
-
-            return new Result(
-                null,
-                Result::CODE_FAIL,
-                'Roles could not be found by roleId.'
-            );
-        }
-
         return new Result($roles);
     }
 
@@ -161,15 +141,6 @@ class DoctrineAclRoleDataMapper
     {
         $roles = $this->getEntityManager()->getRepository($this->getEntityClass())
             ->findBy(array('parentRoleId' => $parentRoleId));
-
-        if (empty($roles)) {
-
-            return new Result(
-                null,
-                Result::CODE_FAIL,
-                'Roles could not be found by parentRoleId.'
-            );
-        }
 
         return new Result($roles);
     }
@@ -206,18 +177,6 @@ class DoctrineAclRoleDataMapper
         $result = $this->getValidInstance($aclRole);
 
         $aclRole = $result->getData();
-        $id = $aclRole->getId();
-
-        if (!empty($id)) {
-
-            $result = $this->fetchById($id);
-
-            if ($result->isSuccess()) {
-
-                return $result;
-            }
-        }
-
         $roleId = $aclRole->getRoleId();
 
         if (!empty($roleId)) {
@@ -227,7 +186,11 @@ class DoctrineAclRoleDataMapper
             return $result;
         }
 
-        return new Result(null, Result::CODE_FAIL, 'Acl Role could not be read.');
+        return new Result(
+            null,
+            Result::CODE_FAIL,
+            'Acl Role could not be read.'
+        );
     }
 
     /**
@@ -239,7 +202,12 @@ class DoctrineAclRoleDataMapper
      */
     public function update(AclRole $aclRole)
     {
-        $result = $this->getValidInstance($aclRole);
+        $result = $this->read($aclRole);
+
+        if(!$result->isSuccess()){
+
+            return $result;
+        }
 
         $aclRole = $result->getData();
 
@@ -249,7 +217,8 @@ class DoctrineAclRoleDataMapper
         // @todo validate action
         return new Result(
             null,
-            Result::CODE_SUCCESS
+            Result::CODE_SUCCESS,
+            'Successfully updated ' . $aclRole->getRoleId()
         );
     }
 
@@ -262,17 +231,61 @@ class DoctrineAclRoleDataMapper
      */
     public function delete(AclRole $aclRole)
     {
-        $result = $this->getValidInstance($aclRole);
+        $result = $this->read($aclRole);
+
+        if(!$result->isSuccess()){
+
+            return $result;
+        }
 
         $aclRole = $result->getData();
+
+        $aclRoleId = $aclRole->getRoleId();
+        $parentRoleId = $aclRole->getParentRoleId();
+
+        $result = $this->fetchByParentRoleId($aclRoleId);
+
+        if(!$result->isSuccess()){
+
+            $result->setMessage(
+                'Failed to find child roles for  ' .
+                $aclRole->getRoleId() . '.'
+            );
+            return $result;
+        }
+
+        $childRoles = $result->getData();
+
+        foreach($childRoles as $childRole){
+
+            $childRole->setParentRoleId($parentRoleId);
+            $childResult = $this->update($childRole);
+
+            if(!$childResult->isSuccess()){
+
+                $result->setCode(Result::CODE_FAIL);
+                $result->setMessage($childResult->getMessage());
+            }
+        }
+
+        if(!$result->isSuccess()){
+
+            $result->setMessage(
+                'Failed to update child roles for delete of ' .
+                $aclRole->getRoleId() . '.'
+            );
+            return $result;
+        }
 
         $this->getEntityManager()->remove($aclRole);
         $this->getEntityManager()->flush();
 
-        // @todo validate action
+        // @todo validate action results
+
         return new Result(
             null,
-            Result::CODE_SUCCESS
+            Result::CODE_SUCCESS,
+            'Successfully deleted ' . $aclRoleId
         );
     }
